allow passing *this* to zeta(-created) functions
until now, use of zeta with object methods was problematic,
DOM event handlers were basically impossible; e. g.

    var o = {};
    var f = function () { return this; };
    var g = function () { return ... };
    o.m = compose(f, g);
    assertEquals(false, o == o.m());

in the example above, f's and g's *this* would have been
"stolen" by zeta, fixing it to "the global object" in the
case of f, or in the case of g, the function itself (because
of zeta's apply), because the function returned from compose
simply did

    return f(apply(g, arguments))

functions listed below are those whose text has changed, but
the impact is broader that that: affected functions include
collect and spread, conjoin and disjoin, use1st and use2nd,
etc.

other parts of zeta could probably benefit from this change
as well, further investigation is in order.  follow-ups on
as-needed base.

* src/base.js

  fix apply, bind1st, bind2nd, compose, bind

* src/function.js

  fix __ZETA__shortcircuit, ifte

* tests/function.js

  new tests