Makefile: `ZETA_JSH` defaults to `node`
it's about time.

NodeJS dominates the market with (by a large margin) largest
mindshare, is widely available, including Windows (the project
distributes Windows binaries and installers), and they have
terrific velocity with no sign of running out of fuel.

v8, the JS engine by Google, releases frequently, and NodeJS
frequently updates its internal, patched copy.  it's fast and
getting faster every fortnight.

in comparison, SpiderMonkey is distributed in source only
which is a hurdle for Windows users, releases are far between,
so i'm not holding my breath for improvement.

also, i'm experimenting a bit with CoffeeScript, and that is
a NodeJS application, so this is another factor.